Tests/LibWeb/Text/expected/Fetch/request-constructor-does-not-throw-on-null-body-with-get-or-head-request.txt

Successfully created GET request with body set to null
Successfully created HEAD request with body set to null
Successfully started GET fetch with body set to null
Successfully started HEAD fetch with body set to null
